backend/pyproject.toml
[project]
name = "andys-daily-factoids-backend"
version = "0.1.0"
description = "Django backend for Andy's Daily Factoids"
requires-python = ">=3.10"
dependencies = [
    "Django>=5.0,<6.0",
    "djangorestframework>=3.15,<4.0",
    "django-ratelimit>=4.1,<5.0",
    "psycopg[binary]>=3.1,<4.0",
    "httpx>=0.27,<0.28",
    "stripe>=11.0,<12.0",
    "django-cors-headers>=4.3,<5.0",
    "whitenoise>=6.6,<7.0",
    "structlog>=24.1,<25.0",
    "langgraph>=0.1.0",
    "posthog>=3.4,<4.0",
    "pydantic>=2.6,<3.0",
    "dj-database-url>=2.1,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1,<9.0",
    "pytest-django>=4.8,<5.0",
    "coverage[toml]>=7.4,<8.0",
    "factory-boy>=3.3,<4.0",
    "ruff>=0.4.5",
    "mypy>=1.9,<2.0",
]

[tool.uv]
python = "3.10"
